Updated Code

+++ Default/Areas/Administration/Models/SegmentsListItemModel.cs
@@ -8,9 +8,9 @@ namespace Default.Areas.Administration.Models
         public string Name { get; set; }
         public string Reach { get; set; }
         public List<SegmentOptionModel> AllOptions { get; set; }
-        public float Past30DaysGrowthInPct { get; set; }
-        public float Past60DaysGrowthInPct { get; set; }
-        public float Past90DaysGrowthInPct { get; set; }
+        public double Past30DaysGrowthInPct { get; set; }
+        public double Past60DaysGrowthInPct { get; set; }
+        public double Past90DaysGrowthInPct { get; set; }
         public List<string> BasicOptions { get; set; }
     }
 }
